my_api_agent/requirements.txt
pocketflow
requests
PyYAML
openai # Or your preferred LLM client library 